requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
mysqlclient
firebase-admin
alembic
python-dotenv
PyMySQL